[homework-9-final]
[homework-9-final]

+++ src/main/resources/liquibase/changelogs/homework-2/sql/01.create-users_profiles-table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "users_profiles" (
-	"id" serial NOT NULL,
+	"id" bigint NOT NULL,
 	"name" varchar(255) NOT NULL,
 	"surname" varchar(255) NOT NULL,
 	"email" varchar(255) NOT NULL,
