+++ src/main/resources/liquibase/changelogs/homework-2/sql/08.create-users_auth-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "users_auth" (
-	"id" bigint NOT NULL,
-	"role_id" bigint NOT NULL,
+	"id" serial NOT NULL,
+	"role_id" serial NOT NULL,
 	"login" varchar(255) NOT NULL UNIQUE,
 	"password" varchar(255) NOT NULL,
 	CONSTRAINT "users_auth_pk" PRIMARY KEY ("id")
